+++ lib/clixon/clixon_xml_diff.h
@@ -39,8 +39,30 @@
 #define _CLIXON_XML_DIFF_H_
 
 /*
+ * Types
+ */
+/*! Rebase struct, pointers to existing objects */
+struct diff_rebase{
+    cxobj  **dr_addparent;  /* Parent to addsub copy of child object */
+    cxobj  **dr_addchild;   /* Object to copy */
+    size_t   dr_addlen;
+    cxobj  **dr_remove;     /* Object to purge */
+    size_t   dr_removelen;
+    cxobj  **dr_changefrom; /* Object source body (value has changed) */
+    cxobj  **dr_changeto;   /* Object destination body */
+    size_t   dr_changelen;
+};
+typedef struct diff_rebase diff_rebase_t;
+
+/*
  * Prototypes
  */
-int xml_rebase_check(clixon_handle h, cxobj *x0, cxobj *x1, cxobj *x2, int *conflict);
+diff_rebase_t *diff_rebase_new(void);
+int            diff_rebase_free(diff_rebase_t *dr);
+int            diff_rebase_exec(diff_rebase_t *dr);
+int            xml_rebase(clixon_handle h, cxobj *x0, cxobj *x1, cxobj *x2,
+                          int *conflict, diff_rebase_t *dr);
+// XXX Backward compatible for branch privcand
+#define xml_rebase_check(h, x0, x1, x2, c) xml_rebase((h), (x0), (x1), (x2), (c), NULL)
 
 #endif  /* _CLIXON_XML_DIFF_H_ */